// generated/src/app/api-models/group-panel.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class GroupPanel {
  seqGroupPanel: number;
  seqGroupId: number;
  panelId: string;
  planCode: string;
  effectiveDate: Date;
  termDate: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}